comments and posts pagination

+++ comment/views.py
@@ -9,45 +9,46 @@ from rest_framework.response import Response
 from rest_framework import status
 from post.models import Post
 from django.db.models import F
+from comment import serializers
 
 # Create your views here.
 
-class CommentList(generics.ListAPIView):
+class CommentList(generics.ListCreateAPIView):
     queryset=Comment.objects.all()
     serializer_class=CommentSerializer
-    permission_classes=[customPermissons.ObjectPermissions]
-    
- 
-    
-@api_view(['POST',"GET"])
-@permission_classes([customPermissons.POSTCommentPermissions])
-def comment_list(request):
-    if request.method == 'POST':
-        serializer = CommentSerializer(data=request.data)
+    permission_classes=[customPermissons.CommentPermissions]
+    def post(self, request, *args, **kwargs):
+        serializer = serializers.CommentSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
             print(serializer.data)
             post = Post.objects.filter(uid = serializer.data['post_id']).update(comment_count=F('comment_count')+1)
             return Response(serializer.data,status=status.HTTP_201_CREATED)
         return Response(serializer.errors,status=status.HTTP_400_BAD_REQUEST)
-    if request.method == "GET":
+    def get(self, request, *args, **kwargs):
         comments = Comment.objects.all()
-        serializers = CommentSerializer(comments,many=True)
+        serializer = serializers.CommentNestedPostSerializer(comments,many=True)
         print('swer5')
-        return Response(serializers.data,status=status.HTTP_200_OK)
-
-class CommentDetail(generics.RetrieveUpdateAPIView):
+        return Response(serializer.data,status=status.HTTP_200_OK)
+    
+class CommentDetail(generics.RetrieveUpdateDestroyAPIView):
     queryset=Comment.objects.all()
-    serializer_class=CommentSerializer
+    serializer_class=serializers.CommentSerializer
     lookup_field='uid'
-    permission_classes=[customPermissons.ObjectCommentPermissions]
+    permission_classes=[customPermissons.CommentPermissions]
     
+    def delete(self, request,uid, *args, **kwargs):
+        comment = Comment.objects.get(pk=uid)
+        self.check_object_permissions(self.request,comment)
+        print(request.user,request.user.role)
+        serializer = CommentSerializer(comment)
+        post = Post.objects.filter(uid = serializer.data['post_id']).update(comment_count=F('comment_count')-1)
+        comment.delete()
+        return Response({'success':'True'})
     
-@api_view(['DELETE'])
-@permission_classes([customPermissons.ObjectCommentPermissions])
-def delete_comment(request,uid):
-    comment = Comment.objects.get(pk=uid)
-    serializer = CommentSerializer(comment)
-    post = Post.objects.filter(uid = serializer.data['post_id']).update(comment_count=F('comment_count')-1)
-    comment.delete()
-    return Response({'success':'True'})
+
+@api_view(['GET'])
+def comments_by_post(request,uid):
+    comments = Comment.objects.filter(post_id = uid,parent__isnull=True)
+    serializer = serializers.CommentNestedPostSerializer(comments,many=True)
+    return Response(serializer.data,status=status.HTTP_200_OK)
